Добавлен внутриигровой UI отображающий ХП и Щит.

+++ Assets/SpaceShooter/ScriptableObjects/Scripts/Stats/PlayerStatsObject.cs
@@ -3,24 +3,24 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerStats", menuName = "ScriptableObjects/Stats/PlayerStats")]
 public class PlayerStatsObject : StatsObject
 {
-    [SerializeField] private float maxHealthValue = 10f;
-    public float MaxHealthValue
+    [SerializeField] private int maxHealthValue = 10;
+    public int MaxHealthValue
     {
         get { return maxHealthValue; }
         set { maxHealthValue = value; }
     }
     public int HealthLvl = 0;
 
-    [SerializeField] private float maxShieldValue = 10f;
-    public float MaxShieldValue
+    [SerializeField] private int maxShieldValue = 10;
+    public int MaxShieldValue
     {
         get { return maxShieldValue; }
         set { maxShieldValue = value; }
     }
     public int ShieldLvl = 0;
 
-    [SerializeField] private float baseSpeed = 20f;
-    public float Speed
+    [SerializeField] private int baseSpeed = 20;
+    public int Speed
     {
         get { return baseSpeed; }
         set { baseSpeed = value; }
